+++ example/lib/simple_table_page.dart
@@ -2,8 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:table_sticky_headers/table_sticky_headers.dart';
 
 class SimpleTablePage extends StatelessWidget {
-  SimpleTablePage(
-      {required this.data, required this.titleColumn, required this.titleRow});
+  SimpleTablePage({
+    required this.data,
+    required this.titleColumn,
+    required this.titleRow,
+  });
 
   final List<List<String>> data;
   final List<String> titleColumn;
@@ -23,6 +26,7 @@ class SimpleTablePage extends StatelessWidget {
         rowsTitleBuilder: (i) => Text(titleRow[i]),
         contentCellBuilder: (i, j) => Text(data[i][j]),
         legendCell: Text('Sticky Legend'),
+        // offset: 300,
       ),
     );
   }