+++ lib/src/custom_properties/custom_scroll_physics.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/widgets.dart';
+
+class CustomScrollPhysics {
+  CustomScrollPhysics({
+    this.stickyRow,
+    this.stickyColumn,
+    this.contentVertical,
+    this.contentHorizontal,
+  });
+
+  final ScrollPhysics? stickyRow;
+  final ScrollPhysics? stickyColumn;
+  final ScrollPhysics? contentVertical;
+  final ScrollPhysics? contentHorizontal;
+}