+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Release Date
-* 05.11.2020
+* 31.08.2022
+
+## [2.0.3]
+* Added ability to jump to index programmatically
+* Fixed problem with 'too sticky' headers where table gets jumping to 0,0 upon content update
+* Cleanup & code organization in SRC
 
 ## [2.0.2]
 * Added RTL support (thanks @Islam-alshiki)